+++ packages/gallery/src/components/gallery/leanGallery/isEligible.js
@@ -1,11 +1,7 @@
 import consts from '../../../common/constants/index';
-import { addLayoutStyles } from '../../helpers/layoutHelper';
-// TODO: DONT_SHOW and allowHover are not supported anymore, ask guy
 //example: http://pro-gallery.surge.sh/?titlePlacement=DONT_SHOW&itemClick=nothing&allowTitle=true&allowHover=false&galleryLayout=2&allowLeanGallery=true
 
-export default ({items, styles, options}) => {
-
-    styles = {...styles, ...options};
+export default ({items, styles}) => {
 
     const allowLeanGallery = !!styles.allowLeanGallery;
     
@@ -22,15 +18,14 @@ export default ({items, styles, options}) => {
         return false;
       }
     }
-    const fullStyles = addLayoutStyles(styles);
-    for (const [styleParam, value] of Object.entries(fullStyles)) {
-      if (!isValidStyleParam(styleParam, value)) {
+    for (const [styleParam, value] of Object.entries(styles)) {
+      if (!isValidStyleParam(styleParam, value, styles)) {
         console.log(`[LEAN GALLERY] NOT ALLOWED - invalid styleParam`, styleParam, value);
         return false;
       }
     }
 
-    console.log(`[LEAN GALLERY] ALLOWED!`, fullStyles);
+    console.log(`[LEAN GALLERY] ALLOWED!`, styles);
     return true;
     
 }
@@ -44,12 +39,14 @@ const isImage = item => {
   return isImageItem;
 }
 
-const isValidStyleParam = (styleParam, value) => {
+const isValidStyleParam = (styleParam, value, allStyles) => {
   if (typeof handledStyleParams[styleParam] !== 'undefined') return true;
   if (typeof ignoredStyleParams[styleParam] !== 'undefined') return true;
   if (typeof fixedStyleParams[styleParam] !== 'undefined') {
     const sp = fixedStyleParams[styleParam];
-    if (sp && sp.length > 0) {
+    if (sp && typeof sp === 'function') {
+      return sp(allStyles);
+    } else if (sp && sp.length > 0) {
       return sp.includes(value);
     } else {
       return sp === value;
@@ -62,6 +59,9 @@ const isValidStyleParam = (styleParam, value) => {
 //these styles can get any value, the lean gallery will handle them
 const handledStyleParams = { 
   numberOfImagesPerRow: 3,
+  gallerySizeType: 'smart',
+  gallerySizeRatio: 1,
+  gallerySizePx: 300,
   gallerySize: 30,
   cubeType: 'fill',
   cubeRatio: 1,
@@ -72,6 +72,7 @@ const handledStyleParams = {
   itemBorderWidth: 0,
   itemBorderRadius: 0,
   imageQuality: 90,
+  textBoxHeight: 200,
 };
 
 //these params are not relevant when a lean gallery is rendered - the fixed styles will override them
@@ -124,7 +125,6 @@ const ignoredStyleParams = {
   galleryThumbnailsAlignment: 'bottom',
   thumbnailSize: 250,
   magicLayoutSeed: 1,
-  textBoxHeight: 200,
   textImageSpace: 10,
   textBoxBorderRadius: 0,
   textBoxBorderWidth: 0,
@@ -157,7 +157,6 @@ const ignoredStyleParams = {
   loadMoreAmount: 'all',
   addToCartBorderWidth: 1,
   imageLoadingMode: consts.loadingMode.BLUR,
-  calculateTextBoxHeightMode: consts.calculationOptions.AUTOMATIC,
   hoveringBehaviour: consts.infoBehaviourOnHover.APPEARS,
   expandAnimation: consts.expandAnimations.NO_EFFECT,
   imageHoverAnimation: consts.imageHoverAnimations.NO_EFFECT,
@@ -190,11 +189,12 @@ const fixedStyleParams = {
   mobilePanorama: false,
   enableInfiniteScroll: true,
   useCustomButton: false,
-  externalInfoHeight: 0,
   bottomInfoHeight: 0,
+  externalInfoHeight: 0,
   itemEnableShadow: false,
   usmToggle: false,
-  itemClick: consts.itemClick.NOTHING,
+  itemClick: [consts.itemClick.NOTHING, consts.itemClick.LINK, consts.itemClick.FULLSCREEN, consts.itemClick.EXPAND],
   scrollAnimation: consts.scrollAnimations.NO_EFFECT,
-  titlePlacement: consts.placements.SHOW_ON_HOVER,
+  titlePlacement: [consts.placements.SHOW_ABOVE, consts.placements.SHOW_BELOW, consts.placements.DONT_SHOW],
+  calculateTextBoxHeightMode: sp => sp.titlePlacement === consts.placements.DONT_SHOW || sp.calculateTextBoxHeightMode ===consts.calculationOptions.MANUAL,
 };