+++ packages/gallery/src/components/item/itemView.js
@@ -778,15 +778,16 @@ class ItemView extends GalleryComponent {
   }
 
   getItemContainerStyles() {
-    const { styleParams, linkData } = this.props;
+    const { styleParams, linkData, linkUrl } = this.props;
     const containerStyleByStyleParams = getContainerStyle(styleParams);
+    const itemDoesntHaveLink = linkData.type === undefined && (linkUrl === undefined || linkUrl === ''); //when itemClick is 'link' but no link was added to this specific item
     const itemStyles = {
       overflowY: styleParams.isSlideshow ? 'visible' : 'hidden',
       position: 'absolute',
       bottom: 'auto',
       margin: styleParams.oneRow ? styleParams.imageMargin + 'px' : 0,
       cursor: styleParams.itemClick === CLICK_ACTIONS.NOTHING ||
-        (styleParams.itemClick === CLICK_ACTIONS.LINK && linkData.type === undefined) //when itemClick is 'link' but no link was added to this specific item
+      (styleParams.itemClick === CLICK_ACTIONS.LINK && itemDoesntHaveLink)
         ? 'default'
         : 'pointer'
     };