+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [v1.6.12](https://pro-gallery-1-6-12.surge.sh) (19/01/2020)
+ 
+#### GALLERY
+ -  remove arrows from first and last items
+
+---
+## [v1.6.11](https://pro-gallery-1-6-11.surge.sh) (19/01/2020)
+ 
+#### GALLERY
+ -  fix: items will be clickable when itemClick is link and there is a dynamic link on those items.
+
+---
 ## [v1.6.10](https://pro-gallery-1-6-10.surge.sh) (16/01/2020)
  
 #### GALLERY